+++ fastcrypto-zkp/src/bn254/unit_tests/zk_login_tests.rs
@@ -141,11 +141,11 @@ fn test_verify_zk_login_google() {
 }
 
 #[test]
-fn test_verify_zk_login_facebook() {
+fn test_verify_zk_login_twitch() {
     let eph_pubkey = big_int_str_to_bytes(
         "84029355920633174015103288781128426107680789454168570548782290541079926444544",
     );
-    let zklogin_inputs = ZkLoginInputs::from_json("{\"proof_points\":{\"pi_a\":[\"1506355985430918561986621304944206628720275836415696432527226322011786292469\",\"613455373800062466065258859046532491698424798267244949352433118740950052820\",\"1\"],\"pi_b\":[[\"13432605760939024875746197498117682238269308826467106768856049869617495266990\",\"1832067857097543048080313900700732174583103404004828871583233404786491016938\"],[\"12774506805709477874931076799272413990755625719529677477582407925650184326419\",\"9920407680353330309465070375503195109002963389725451415597959332611544046474\"],[\"1\",\"0\"]],\"pi_c\":[\"21460871700073370465890462779698742952227192461460518900819221976004900851129\",\"16417973205725808622705510292771877448246769946278068014196442886693189946200\",\"1\"]},\"address_seed\":\"8309251037759855865804524144086603991988043161256515070528758422897128118794\",\"claims\":[{\"name\":\"iss\",\"value_base64\":\"wiaXNzIjoiaHR0cHM6Ly9pZC50d2l0Y2gudHYvb2F1dGgyIiw\",\"index_mod_4\":2},{\"name\":\"aud\",\"value_base64\":\"yJhdWQiOiJyczFiaDA2NWk5eWE0eWR2aWZpeGw0a3NzMHVocHQiLC\",\"index_mod_4\":1}],\"header_base64\":\"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IjEifQ\"}").unwrap().init().unwrap();
+    let zklogin_inputs = ZkLoginInputs::from_json("{\"proof_points\":{\"pi_a\":[\"13091564805270242091988535637094928145526962426048248457544204014146120848061\",\"16720184877774297257067413263271993886875650921635894599271213076438028149121\",\"1\"],\"pi_b\":[[\"10315437641058147137376024432028897182269999873576264368551692657896041757048\",\"1508628913652029422583835336497010615341442910432485486157510667573899391209\"],[\"2615204984444202339450727742801445002386326248143849542511762200103437621802\",\"5849486215298305357746845827664545444623657491883886014999288709074461992087\"],[\"1\",\"0\"]],\"pi_c\":[\"14195277018213067909432749083395103683618329379082932858378933760763589481512\",\"15923172920578108290608652430421506204931592172775947506845238271513931833500\",\"1\"]},\"address_seed\":\"8309251037759855865804524144086603991988043161256515070528758422897128118794\",\"claims\":[{\"name\":\"iss\",\"value_base64\":\"wiaXNzIjoiaHR0cHM6Ly9pZC50d2l0Y2gudHYvb2F1dGgyIiw\",\"index_mod_4\":2},{\"name\":\"aud\",\"value_base64\":\"yJhdWQiOiJyczFiaDA2NWk5eWE0eWR2aWZpeGw0a3NzMHVocHQiLC\",\"index_mod_4\":1}],\"header_base64\":\"eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IjEifQ\"}").unwrap().init().unwrap();
     assert_eq!(zklogin_inputs.get_kid(), "1".to_string());
     assert_eq!(
         zklogin_inputs.get_iss(),
@@ -175,8 +175,13 @@ fn test_verify_zk_login_facebook() {
         n: "6lq9MQ-q6hcxr7kOUp-tHlHtdcDsVLwVIw13iXUCvuDOeCi0VSuxCCUY6UmMjy53dX00ih2E4Y4UvlrmmurK0eG26b-HMNNAvCGsVXHU3RcRhVoHDaOwHwU72j7bpHn9XbP3Q3jebX6KIfNbei2MiR0Wyb8RZHE-aZhRYO8_-k9G2GycTpvc-2GBsP8VHLUKKfAs2B6sW3q3ymU6M0L-cFXkZ9fHkn9ejs-sqZPhMJxtBPBxoUIUQFTgv4VXTSv914f_YkNw-EjuwbgwXMvpyr06EyfImxHoxsZkFYB-qBYHtaMxTnFsZBr6fn8Ha2JqT1hoP7Z5r5wxDu3GQhKkHw".to_string(),
         alg: "RS256".to_string(),
     });
-    let res = verify_zk_login(&zklogin_inputs, 10000, &eph_pubkey, &map, Environment::Test);
+    let res = verify_zk_login(&zklogin_inputs, 10, &eph_pubkey, &map, Environment::Test);
     assert!(res.is_ok());
+}
+
+#[test]
+fn test_verify_zk_login_facebook() {
+    // TODO
 }
 
 #[test]
